Add `Join` class.
